Reorganize description box rendering
Added a comment on the addToProjectUI to give a clearer description on
what the function does along with the DOM targeting for the project
title under it. Added another comment on a possible option for the
taskName variable (going to ul/li method instead of the current <p>
element).

Moved the conditional statement for the description box rendering to
the correct place so that the description elements render at the bottom
of the task like it should.