Refactor render & store functions & button logic
Added the 'trim' method to all variables and any other code having to do
with the title of a project or task. This was added to eliminate any
possible whitespace issues.

Moved all the render function calls that were inside the storeTask and
storeProject functions in the tasks and projects modules over to the
Confirm button logic index button module. This just makes more sense
since we want to keep UI rendering separate from any application logic
already in those modules.

Removed a duplicate line of code that was calling the readProject method
inside of storeTask, then calling it again under a declared variable
that later function as a condition for an 'if' statement under the same
storeTask function. Basically we were calling it twice when we didn't
need to.

Tested all changes and the app still functions correctly in it's current
state.