Add Update and Delete icon buttons to tasks
Added in the same style Update and Delete icon buttons that render with
each project to all tasks as well. Resized all Update and Delete icon
buttons as they were a bit too big (task buttons are smaller than the
project buttons). Will reorganize placement of Update and Delete buttons
in the future, particularly for projects. Added color to each of the
icon buttons as well.

Changed the values of the Priority options menu in the HTML as they were
showing the color of the Priority as text and not the text itself
(Urgent, Important, etc.).