Refactor target project for checkbox in renderTask
Added a line of code in an attempt to represent the current task being
rendered in the renderTask function. This is for the purpose of the
checkbox elements being rendered. The 'id' and 'for' attributes were
not showing up correctly for the target project. Changed the attributes
to reflect the new variable for the target project being declared. Also
changed the 'isChecked' variable to reflect this as well.

Gave the checkboxes some more CSS styling.