Add checkbox functionality to index module
Added checkbox functionality to the index module. Now the Add Project
and Update Project buttons don't work. Commented out the functionality
for now. Variable that was supposed to target the current task in the
renderTask function isn't working now either. Made a note of this but
leaving the code in for now.