Resolve TypeErrors related to checkbox rendering
Added a getTask function expression and used that for all the id and
type attributes for the checkbox elements in renderTask function. The
prior attributes and attempts to get the checkboxes to render while
maintaining app functionality were causing several TypeErrors. The
getTask function resolves all of those errors. CSS styling of the
checkboxes are confirmed to show in the UI correctly.

Commented out the imported functions in the projects module as I doubt
anything will need to use functions from the tasks module there.

Next step is trying to figure out the click eventListener for the
checkboxes. The current commented out code in the index module could
still potentially break functionality in the app like it did before.