Refactor checkbox functionality in index module
First reinstated the data-key set attribute inside the UI module in an
attempt to get the previous checkbox functionality to work, but to no
avail. Left in the setAttribute for the data-key to be used by the new
function. The new checkbox logic is a forEach method. It is the only
solution so far that doesn't throw a TypeError, but it doesn't actually
work (toggle the checkboxes).

The app runs through the event listeners in the main area itself that
were built for the project UI functionality, then skips the checkbox
logic altogether. Added in some comments regarding this in the index
module and general comments on future "bugs" that need to be tackled in
the UI module.