Modify search bar styling & debug test checkboxes
Made an attempt to debug the checkboxes in the index module. Currently,
as per previous commit, when the checkbox is clicked, the app runs
through the Update and Delete Project button logic in the index module.
The browser then sees that the checkbox doesn't match any condition and
so completes both functions with nothing happening in the UI. It never
gets to the checkbox function in the index module.

The attempt was to switch the event listeners for the Projects from
the <main> area to the document itself based on a previous comment I
left as an option for those functions. But this makes it so no Update
or Delete Project icon buttons work. Added some code to switch the
checked status in the target task from false to true in the checkbox
function.

Added beginning whitespace to the placeholder tag in the search section
of the HTML file to move that text over so it isn't so close to the
edge of the search bar.

Updated the styling and font of the search bar itself. The bar is now
larger with some inset border styling, giving it some more thickness.
Added 'appearance: none' to account for some Safari browser styling
restrictions.