Modify checkbox forEach method in index module
Added a <main> area event listener to the checkbox forEach method in the
index module. This solves the bug of the browser completely skipping
the method and going straight to running the other two mainArea event
listeners (Update & Delete Project). However, the browser runs through
the listener twice and none of the code in the forEach method is ran.
This likely means something is wrong with the forEach method itself.
Added some comments with possible issues and possible solutions/next
steps.

Will also likely need to add <main> area even listeners to everything
else that's remaining in order to get those things to work, namely
Update & Delete Tasks buttons.