Refactor checkbox method in index module
Refactored the checkbox method inside the index module in an attempt to
resolve the checkboxes not working as intended (no strikethroughs on
any text, nothing showing the task is complete in the UI or the array,
or the checkbox showing as "checked"). It is closer but not quite
there yet. tagName method has helped and the browser actually runs
through the method correctly but nothing actually happens on the UI.

Commented a TypeError under the readTask function in the tasks module
to reflect that it doesn't run correctly with the current tagName
conditional inside the checkbox method.

Replaced all tags in the renderTask function that had the task id number
as their content to the title text entered into the form directly for
the time being.

Added and removed comments throughout (new TODO items, OPTIONS that we
can pursue, notable error messages etc.)