Add Update Task button functionality & comments
Added new functionality to the Update Task icon buttons in the index
module similar to how the Update Project icon buttons work. Modified
the Update button functionality in the Update Task form as well. Will
test out the changes in the near future.

Added comments regarding alternate method of running renderProject
which would get the project info directly from the array itself and
not the form data entered by the user. Comments state there are pros
and cons to both. Leaving the function the way it is (form data) for
the time being.