Refactor renderTask function to resolve TypeError
Refactored the renderTask function due to a TypeError showing up that
couldn't read the ID number of the task due to it always showing as
undefined in the DOM. This was because getTask was not being called,
then later used in the correct manner. Resolved this issue and will be
testing the changes very soon.

The Update and Delete Task button logic in the index module was also
updated to reflect the renderTask changes. The buttons should work now
but will test them out very soon.