Partially revert renderTask function with comments
Reverted the section of the renderTask function back to the previous
version, although slightly modified. When declaring a separate variable
to use for calling 'getTask', we get a TypeError regarding the 'find'
method under getTask and we don't even get out of the form we're in.
Currently, the new conditional statement calls getTask as a parameter
with the old attributes being shown again (still undefined in the DOM).
So we are back to where we started.

Added some comments regarding invalid form control that isn't focusable
in the project form section of the HTML as well as a comment regarding
the task not being found in the array at all when trying to update it
via the Update Task button functionality. Delete Task button appears to
not do anything at all, but will test again soon.