Comment out an attempted solution to TypeErrors
Attempted to use a solution that directly grabbed the task from the
nested tasks array in the app logic to use the ID number for rendered
attributes instead of trying to use what was entered into the form and
matching the ID number in the array with the form entry. This did not
work unfortunately, at least not on the first attempt.