Add console log debug code to renderTask
Added a console log debugger to the renderTask function to see whether
or not 'tasks' imported from the task.js module was being read as an
array. It's not. Will step through the code line by line in DevTools
to see exactly how the code is running and how to implement getting
the ID from the task nested inside the target project.