Refactor renderProject resolving attribute errors
Refactored the renderProject and was finally able to resolve the errors
of attributes showing undefined instead of the ID number. After many
attempts, I realized I had no way of targeting the correct project that
the task was under for the UI portion (I had it for the app logic side).
Declared a variable to target the project in the renderTask function,
then went with a forEach method that allows the app to step through each
task within that project. Function now renders based on a condition of
whether the task title matches the title entered in the form AFTER the
correct project is identified (whereas before there was no project to
identify). ID numbers of the tasks now show correctly as attribute
values for all relevant elements/nodes.

Updated some comments and TODOs based on the error fix.