Change header styling and footer positioning
Changed the box shadow styling of the header to a drop shadow to see if
it will look better in the UI. Will test different colors of the drop
shadow in a future update. Added absolute positioning back to the footer
with some additional CSS to see if that will resolve the issue of the
footer not sticking to the bottom of the page.