Resolve null error for Update Task button logic
The "Update Task" icon button logic was throwing null for the class of
the DOM element we were targeting for a variable. It turned out to be
a simple typo error. The "Update Task" button seems to be working
correctly now and pulling up the "Update Task" form as intended.

A new error is showing up stating the property/value of the task title
is throwing null in the updatedTask object when trying to actually
update the task via click of the Update button inside the actual
"Update Task" form itself.