Modify updatedTask object to resolve errors
Resolved the errors of each property inside the updatedTask object which
is in our Update Task form button logic that were throwing 'null'. The
values of the HTML elements inside the Update Task form section were
never updated to reflect what we had in the JS code. Merely needed to
edit them to match up correctly.

We now have an issue where readTask in the tasks module is throwing a
type error when trying to read the task that's to be updated (coming up
as 'undefined'). Made TODO comments to reflect this.