Refactor readTask and add TODO comments throughout
Refactored readTask to a for...in loop in an attempt to resolve the
"undefined" errors being thrown when trying to update a task in the UI
via update form. App now finishes stepping through the code and gets
back to the UI page, closing the form. However, no changes are made
due to the 'projectTitle' now showing undefined with the way the code in
readTask is written now.

Added several TODO comments throughout reflecting this (in readTask,
updateTask and Update Task button logic in all relevant modules). Also
added TODO comments regarding tackling the "Add to Project" populate
list issue on a task update in both populateProjects function and the
Update Task button logic.