Add a log to check if the tasks array was updating
Added a console log in a debug attempt to check whether or not the task
was updating in the array when the update task form was sent. Have
confirmed it is not updating. Added some comments for the readTask and
updateTask functions to show attempts that have been made to resolve
bugs for help related purposes. It seems like everything points back to
finding the project first, then finding the task by ID number within
that project for anything to work.