Update attributes & reinstate them in the js
Updated the Add to Project attributes for the Update Task form in the
HTML, then reinstated the value related to the Add to Project dropdown
in the Update button logic for tasks. Attempted to add the dropdown
with no other conditional in populateProjects function but it's not
working right now.