Modify event listeners for buttons & add comments
Modified the Update Task and Update Project button event listeners from
the project heading to the main area itself as the buttons were not
working properly otherwise. The only other option would've been to
place all the headings under a forEach method which ends up being even
more code than we already have. The main area makes it simple and
straightforward and also falls in line with the other event listeners,
though makes it more "coupled" (everything relies on the main area for
event listeners).

Added some comments on the readTask, updateTask and readProject
functions for purposes of asking for help.