Modify object naming, comment out code & add notes
Modified the naming of the updatedTask object inside the "Update Task"
button functionality inside of the index module. It is now called
"updatedDetails" as it prevents repetitive naming with other code
related to updating tasks. Changed the property names inside of the
updatedDetails object back to their old names (removed the word "new")
in order to match better with other functions that have the same naming
for those properties/form elements.

Commented out a code attempt for the readTask function inside the tasks
module, but may bring back a modified version of it. We are attempting
to update the task just using the task ID number alone (which the code
recognizes correctly at the moment) instead of trying to grab the
project that it's under. Added a comment next to the taskFind variable
in the updateTask function in the tasks module to describe what was
going on with it when stepping through the code.