Revise code inside of updateTask and readTask
Took a slightly different approach on a previous attempt at getting the
updateTask and readTask functions in the tasks module to update tasks
in the array correctly. They are still not being updated but errors are
no longer being thrown. Currently the old for...in loop is inside of a
second for...in loop that makes everything go through. UI still doesn't
update, but again that is due to renderTask not being set up to render
changes to existing tasks. Still need to get the change in the array
itself to go through in the console.