Modify readTask function to resolve errors
No errors were being thrown when the "Update" button was clicked in the
"Update Task" form. However, the target task in the array within it's
parent project was not being updated. Changed the "for...in" loop in
the readTask function to a "for...of" loop and the task is now being
updated in the array correctly.