Modified render method for updating tasks
Made an initial attempt at correcting the rendering of updated tasks
via a render method inside the Update button logic for the Update Task
form. Currently it updates the title itself but leaves all other child
elements under the task title completely blank. It even removes the
checkbox and update/delete buttons that were under the title. It could
be an issue with how I'm trying to target those children, so it will
either have to be rewritten or a separate render function specifically
for updated tasks would have to be written in the UI module.