Modify Update button logic for Update Task form
Made several attempts to get the task to update in the UI in addition to
the array by changing the rendered elements directly under the "Update"
button logic for the "Update Task" form in the index module. Nothing has
worked so far. Previous forEach method has been removed. The issue seems
to be with how the DOM elements are being selected?

If we can't figure it out, either we add a container in the renderTask
function for all the task elements to live under, then use that to code
up a solution that directly changes those elements in the button logic
(just like we did with "Update Project") or we write up a completely
separate render function in the UI module just for updated tasks.