Refactor removeTask function and add index comments
Refactored the removeTask function in the tasks module as it wouldn't
run in its current state due to '[project]' not being declared anywhere.
It is now declared in a 'for...of' loop similar to the updateTask
function. Added the previous forEach method that was tried before to
update tasks in the UI as a comment in case its needed later on. It is
placed back where it was under the "Update" button logic for the "Update
Tasks" form in the index module.

Will attempt to reorganize how each task is rendered in the UI module
next. This will involve one or more containers/wrappers for each task.