Refactor renderTask function & edit button styling
Refactored the renderTask function in the UI module to render a wrapper
that contains all the elements of a task (checkbox, buttons, text, etc)
as well as another wrapper inside of the initial task wrapper to contain
all the text elements together. Modified some CSS targeting to maintain
the sizing of the Update Task and Delete Task buttons.