Modify the Update button functions for tasks
Modified the event targeting for the Update button to reflect changes
made in the renderTask function to maintain the functionality of the
button itself. Refactored the "Update" button inside the "Update Task"
form to a modified version of the previous forEach method. The task now
seems to update correctly in the UI after this change. Will test other
update scenarios in the future (involving multiple projects/tasks) to
see if this solution will continue to work.