Modify "Delete Task" button logic and removeTask
Modified the "Delete Task" button logic to target the wrapper instead of
the previous <p> tag to maintain correct functionality in the UI. Tested
out the current and a second variation of removeTask to see if the task
is removed from the array itself. Neither variation worked. Made some
comments regarding not needing the 'for...of' loop specifically since
we already have readTask for that. Correct task does get removed in the
UI however.