Refactor removeTask function & add comments
Refactored the removeTask function in the tasks module in another
attempt to get the task to delete from the array itself, not just the
UI. It worked! Delete Task functionality now works correctly! Removed
the console.log in the "Update" button logic for the "Update Task" form
that was used to check the array afterwards as it's no longer needed.

Added a comment under the "Delete Project" button logic in the index
module concerning the <p> element containing the project title still
remaining after project deletion (the element would just be empty with
no text). Related to TODO comment for the button logic. Will test in the
future.

Added a comment in the UI module regarding the sidebar project buttons
in the renderProject function. An option to repurpose the sidebar
buttons.

Removed other comments throughout that were no longer needed (old notes,
old options & code attempts).