Refactor populateProjects function & edit comments
Refactored the populateProjects function as it kept duplicating projects
in the dropdown menu of the Update Task form that already existed.
Changed where the task container was appended in the renderTask
function to better organize the DOM. Changed to a header element for the
project title inside of each function responsible for rendering the
projects. Edited some comments throughout all modules.