Refactor toggle task functions and UI checkboxes
Refactored the toggleTaskChecked function in the tasks module to use
the readTask function to find the correct task to toggle the checked
status on instead of hard-coding the 'for...of' loop into the function.
The readTask function has that loop already & for reasons like this.
This cleaned up the code in the function a bit and may have even allowed
everything to toggle correctly in both the UI and the array.

Refactored the checkbox button logic to no longer use the forEach
method on all checkbox elements before finding the correct one to
toggle. Removed some console.log lines as they are no longer needed for
debugging their respective functions. Toggling the status of each task
now works as intended.