Implement localStorage & modify access to projects
Implemented code into the localStorage module as a first draft for the
web app to utilize localStorage to "save" user data. Exported the code,
then imported it to all other modules that will likely use at least one
of the functions at some point. Modified access to "appState.myProjects"
by declaring a "projects" variable and using "appState.myProjects" as a
value to that variable. Implemented the variable in all modules that
used "appState.myProjects" in any of their code and made the necessary
edits to use the variable instead in order to maintain functionality of
the code.