Add test code for localStorage feature as comments
Added test code in several places to fully implement the localStorage
feature throughout the app, namely in the CRUD functions. Added and
edited other general comments throughout all modules as well.

Next step is to revert back to the previous code pushed to GitHub as
that is the last working demo of the project. Will leave as much code
in related to localStorage as possible as comments during this process.