Revert back to previous version of app
Reverted code back to the last version of the app that was pushed up to
GitHub plus most of the localStorage code that was initially tested, but
has now been commented out for the time being. Will refactor this code
in a future update in order to get it implemented.