Add and revise comments and fix CSS value spacing
Made some adjustments to the spacing of some values in the CSS. Added a
TODO comment to the renderDefault function in the UI module. Added an
OPTION comment to the search functionality in the index module to try
and add the ability to search by task, not just by project. Adjusted the
placement of a TODO comment in the "Update Task" button functionality
in the index module so it makes better sense in context. Reviewed all
lines of code to make sure everything matches up with the previous code
that was pushed to GitHub. This is the version that will be pushed both
to GitHub and GitHub Pages now that everything is still functioning
correctly.

The localStorage will be added later on.