Modify checkbox styling for toggling tasks
Modified the checkbox styling inside the checkbox button logic in the
index module. Checkboxes now correctly show a checkmark with a green
background color when clicking them to mark the task as complete. They
also go back to their original styling correctly when unchecking them.

Attempted to add functionality to the search bar so it would be able to
search for tasks and not just projects, but was unsuccessful. Made a
comment in the appState of the projects module regarding the myProjects
property. It's a reminder to add getFromStorage when localStorage is
being worked on.