Implement getFromStorage to multiple functions
Implemented getFromStorage() function into multiple render functions in
the UI module for testing. Implemented saveToStorage() in the functions
that needed it in the UI module as well. Tested the changes by reloading
the page, but no projects or tasks show up on page load, not even the
default project like before. There is a TypeError that is being thrown
and I believe it's due to naming conventions clashing with each other
(appState import and appState variable in the page load listener).
Will rename the variable and test it in the future.