Modify DOMContentLoaded and localStorage code
Changed the name of the variable accessing getFromStorage() in the
DOMContentLoaded logic in the index module in an attempt to resolve
errors on page load. Info from any previous session still doesn't load
properly after the change. Switched the localStorage code setup back to
a more modified version of some previous code.