Reorganize saveToStorage calls in all modules
Removed the saveToStorage calls from all render functions in the UI
module as the app was currently calling it twice (once in the app logic
when a project or task rendered and again in the render logic). This
doesn't seem to make sense. Added comments about potentially moving
all saveToStorage calls to the button logic in the index module.

Added several TODO comments and other comments to the localStorage and
index modules regarding how the app was currently functioning on page
load and potentially what to do about those behaviors.

Need to come up with separate rendering functions in the UI module whose
only purpose is to render all the projects and tasks from the previous
session off of localStorage among other things/fixes.