Remove all instances of getFromStorage()
Removed all calls to getFromStorage function from all parts of the UI
bundle for the time being. The goal is to try and create separate
functions for the projects and tasks from previous sessions to be
rendered on using localStorage, not by modifying the existing render
functions. Confirmed that the data-attribute and text content of the
sidebar buttons for projects was still displaying correctly after making
previous changes to the names of those respective nodes.