Add new function that renders previous sessions
Added a new function in the UI module specifically responsible for only
rendering any projects and tasks from the previous user session.
Modified the DOMContentLoaded page load logic in the index module to
call the new function on page load. Currently everything seems to load
correctly but there is an issue with the logic rendering duplicates of
the 'undefined' project we were getting earlier, then loading the
default project, then the other projects and finally a new(?) default
project. It's possible that the order of these projects are reversed
in the localStorage however. Will test this in the future. Implemented
a quick fix in the meantime that removes all 'undefined' projects from
the UI, but the extra default project is still left up.

Attempted to remove the getFromStorage call in the appState object in
the projects module to try and resolve the duplication issue before
adding in the quick fix, but this ended up causing only projects to
load on refresh/new sessions, not tasks (the tasks were still there in
localStorage however). Left the getFromStorage call in for now.

Added a comment in the toggle function of the tasks module regarding
the task being removed from the localStorage array once checked off.
Will research this in the future.

Added a console.log to all localStorage functions to keep track of how
the projects and tasks were changing in the localStorage arrays.