Modify page load hotfix and add TODO comments
Modified DOMContentLoaded page load logic in the index module to include
deleting all duplicate "undefined" sidebar buttons with their respective
duplicate "undefined" projects as a hotfix to the duplication issue for
the time being. Added comments on next steps to solve the duplication
issue above the DOMContentLoaded page load logic.

Added TODO comment in the renderDefault function in the UI module
regarding its for...in loop.

Added TODO comment in UI module for the tasks forEach method under
renderPrevSession to find a way to use the 'task' in the for...in loop
which is being declared but not read right now. Possible that it is also
due to it being a for..in loop and not a for...of loop. TODO comment
added by the for...in loop itself also addresses this.