Refactor page load logic & reorganize console logs
Refactored the page load logic under the DOMContentLoaded functionality
to a modified version of previous code that the app was running. The
previous version's conditional operated on if the localStorage did not
exist, which works better for page load on the app. The prior issue
before was that we had no render function but we do now, which allows
us to use the previous version of the page load logic again. Everything
seems to be in working order but we will need to test all of the CRUD
functions for projects and tasks to make sure everything is working,
rendering and saving properly. These are the next steps.

Removed the console log that checked for the localStorage key in the
saveToStorage function under the localStorage module as it's no longer
needed, nor was useful in the first place.

Moved the console log that checked the projects array after storing a
task from the storeProject function in the projects module over to the
"Confirm" task button logic in the index module. It didn't seem to fire
under storeProject during testing and the console log was under that
button logic in a previous update. Will test adding a project again in
the future to see if it fires and the array is showing the correct
updates.