Remove comments no longer needed after debugging
Managed to test and debug all CRUD functionality for projects and tasks.
Things are mostly working properly. The only issues now are that the
wrong task is deleted from its corresponding tasks array when the task
is under the default "Today" project. It's possible that this is the
case for all projects but will need to test this. On page refresh,
multiple tasks are duplicated to the UI of the default project, but none
of the tasks that were added to other projects are loaded. In terms of
the localStorage, the tasks are NOT duplicated there (so it is purely a
UI rendering issue), but any tasks that were in other projects aren't
in their respective arrays either, meaning they are completely deleted
on refresh. Not sure what the issue is here - how I'm calling on
localStorage? Something else?

Removed all comments containing old code or TODOs that have since been
completed as they are no longer needed. Did keep a little bit of old
code as potential hotfixes for the duplication issue in the UI.
See previous GitHub push if we need to revert.