Resolve duplicate tasks error & add comments
Removed the additional for...in loop for tasks that was nested inside
of the for...of loop for projects in the renderPrevSession function in
the UI module. The for...in loop for tasks wasn't actually doing
anything and we didn't need it in order for the addProjectToUI function
and the forEach method for the targeted tasks to work. After deleting
the additional loop, this solved the duplication bug in the UI. Changed
the for...in loop in the renderDefault to a for...of loop to address the
function reading each project by index # instead of the title itself.
Will test this in a future update.

Added a comment regarding the splice method under the removeTask
function in the tasks module. This is in relation to a previous comment
regarding an error where the first task in a given task array is still
being deleted in the app logic regardless of what task was being
targeted by the user for removal. The UI is still correct however. This
will be the next step along with testing renderDefault on another PC.