Resolve removeTask deletion & checked status bugs
Resolved the issue of removeTask deleting the first task of all projects
instead of just the task that the user targeted in the UI. The UI would
show correctly, but not the arrays or the localStorage, meaning on page
load, the session would not show as the user previously left it. A
conditional that works similar to the one in the search bar function
was added along with the findIndex method since the previous code would
get the entire task object, which doesn't work with splice when deleting
the task (it looks for a number primarily). The new check is if the
task object read earlier doesn't equal -1, then and only then would the
task be deleted.

Found another bug where the checked status of true along with all CSS
styling changes would not persist on page load/refresh. A conditional
statement similar to the one found in the checkbox toggle logic in the
index module was added to the renderPrevSession function in the UI
module to resolve this, which it did.

The project is done...for now.

(Will eventually make it prettier with CSS and work on some remaining
TODO comments that are less crucial as well as some OPTION comments
regarding adding new features or refactoring code)