Modify render and Delete Project button functions
Tested the 'for...of' loop change in the renderDefault function in the
UI module to see if the function would still operate the same way.
Confirmed that the function does work the same. Removed the TODO comment
as well as the OPTION comment related to the function as neither one is
needed any more.

Added context to the TODO comment under the populateProjects function
in the UI module, showing that the TODO is linked to the editTask form
button event listener in the index module.

Added preventDefault to the "Delete Project" button functionality in
the index module. This was done in order to prevent the form from
submitting to the browser's backend. Removed the OPTION comment related
to this as its no longer needed.

Implemented a media query in the CSS that allows the page to toggle
between "light mode" and "dark mode" based on the user's OS or browser
theme. Will test this in the future to make sure it works.